requirements.txt
requests==2.11.1
piexif==1.0.13
Pillow==4.2.1
PyHamcrest==1.9.0
pytest==3.1.0